OV-31: * move timeline to studio bundle
