+++ frontend/src/bundles/studio/components/timeline/subrow.tsx
@@ -1,4 +1,4 @@
-import { Box } from '../../../../common/components/components.js';
+import { Box } from '~/bundles/common/components/components.js';
 
 interface SubrowProperties {
     children: React.ReactNode;
